+++ cfme/tests/test_rest.py
@@ -884,85 +884,6 @@ def test_delete_rates(rest_api, rates, multiple):
             rate.action.delete()
 
 
-@pytest.fixture(scope='function')
-def tenants(request, rest_api):
-    parent = rest_api.collections.tenants.get(name='My Company')
-    data = [{
-        'description': 'test_tenants_{}_{}'.format(_index, fauxfactory.gen_alphanumeric()),
-        'name': 'test_tenants_{}_{}'.format(_index, fauxfactory.gen_alphanumeric()),
-        'divisible': 'true',
-        'use_config_for_attributes': 'false',
-        'parent': {'href': parent.href}
-    } for _index in range(0, 3)]
-
-    tenants = rest_api.collections.tenants.action.create(*data)
-    for tenant in data:
-        wait_for(
-            lambda: rest_api.collections.tenants.find_by(name=tenant.get('name')),
-            num_sec=180,
-            delay=10,
-        )
-
-    @request.addfinalizer
-    def _finished():
-        ids = [tenant.id for tenant in tenants]
-        delete_tenants = [tenant for tenant in rest_api.collections.tenants if tenant.id in ids]
-        if len(delete_tenants) != 0:
-            rest_api.collections.tenants.action.delete(*delete_tenants)
-
-    return tenants
-
-
-@pytest.mark.uncollectif(lambda: version.current_version() < '5.5')
-@pytest.mark.parametrize(
-    "multiple", [False, True],
-    ids=["one_request", "multiple_requests"])
-def test_edit_tenants(rest_api, tenants, multiple):
-    if multiple:
-        new_names = []
-        tenants_data_edited = []
-        for tenant in tenants:
-            new_name = fauxfactory.gen_alphanumeric().lower()
-            new_names.append(new_name)
-            tenant.reload()
-            tenants_data_edited.append({
-                "href": tenant.href,
-                "name": "test_tenants_{}".format(new_name),
-            })
-        rest_api.collections.tenants.action.edit(*tenants_data_edited)
-        for new_name in new_names:
-            wait_for(
-                lambda: rest_api.collections.tenants.find_by(name=new_name),
-                num_sec=180,
-                delay=10,
-            )
-    else:
-        tenant = rest_api.collections.tenants.get(name=tenants[0].name)
-        new_name = 'test_tenant_{}'.format(fauxfactory.gen_alphanumeric().lower())
-        tenant.action.edit(name=new_name)
-        wait_for(
-            lambda: rest_api.collections.tenants.find_by(name=new_name),
-            num_sec=180,
-            delay=10,
-        )
-
-
-@pytest.mark.uncollectif(lambda: version.current_version() < '5.5')
-@pytest.mark.parametrize(
-    "multiple", [False, True],
-    ids=["one_request", "multiple_requests"])
-def test_delete_tenants(rest_api, tenants, multiple):
-    if multiple:
-        rest_api.collections.tenants.action.delete(*tenants)
-        with error.expected("ActiveRecord::RecordNotFound"):
-            rest_api.collections.tenants.action.delete(*tenants)
-    else:
-        tenant = tenants[0]
-        tenant.action.delete()
-        with error.expected("ActiveRecord::RecordNotFound"):
-            tenant.action.delete()
-
-
 @pytest.mark.uncollectif(lambda: version.current_version() < '5.5')
 def test_run_report(rest_api):
     report = rest_api.collections.reports.get(name='VM Disk Usage')
